LayoutTests/http/tests/security/contentSecurityPolicy/inline-style-attribute-blocked-expected.txt

CONSOLE MESSAGE: line 10: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'none'".

PASS
